+++ src/main/java/com/example/team2_be/kakao/util/KakaoClient.java
@@ -1,14 +1,20 @@
-package com.example.team2_be.kakao;
+package com.example.team2_be.kakao.util;
 
+import com.example.team2_be.kakao.DTO.KakaoInfo;
+import com.example.team2_be.kakao.DTO.KakaoToken;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.net.URI;
 
 @FeignClient(name = "kakaoClient", configuration = KakaoFeignConfiguration.class)
 public interface KakaoClient {
+
+    @PostMapping
+    KakaoInfo getInfo(URI baseUrl, @RequestHeader("Authorization") String accessToken);
+
     @PostMapping
     KakaoToken getToken(URI baseUrl, @RequestParam("client_id") String restApiKey,
                         @RequestParam("redirect_uri") String redirectUrl,
